feat: adjust examples/algorithm structure & add label-propagation demo
